refactor: update imports to use startup::AppContext
- Replace context::AppContext with startup::AppContext across routers
- Update account, keys, profiles, and registration modules
- Update auth middleware import path
